+++ editor/tiled/setup.ts
