+++ editor/tiled/tiled_object/MapEditor.ts
@@ -1,5 +1,6 @@
+import {TilesetsView} from './TilesetsView'
 
-interface MapEditor {
+export interface MapEditor {
     // currentBrush : TileMap,//	Get or set the currently used tile brush.
     // currentMapView : MapView,// [read‑only]	Access the current map view.
     tilesetsView : TilesetsView,// [read‑only]	Access the Tilesets view.
