+++ editor/tiled/tiled_object/Point.ts
