+++ editor/tiled/tiled_object/TileLayerEdit.ts
@@ -1,3 +1,4 @@
+import {Tile} from './Tile'
 
 export interface TileLayerEdit {
     setTile: (x : number, y : number, tile : Tile ) => void,
