+++ editor/tiled/tiled_object/TilesetsView.ts
@@ -1,5 +1,7 @@
+import {Tile} from './Tile'
+import {Tileset} from './Tileset'
 
-interface TilesetsView {
+export interface TilesetsView {
     currentTileset : Tileset,//	Access or change the currently displayed tileset.
     selectedTiles : Tile[],//	A list of the tiles that are selected in the current tileset.
 }
