--- tiled/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "TiledExtensionForKitty",
-  "version": "1.0.0",
-  "description": "Adds .kitty file support to the Tiled editor",
-  "private": true,
-  "scripts": {
-    "build": "npx webpack",
-    "install": "npm run build && npm run setup",
-    "dev": "ts-node src/app.ts",
-    "start": "node --unhandled-rejections=strict --inspect=5859 -r ts-node/register ./src/app.ts",
-    "setup": "node --unhandled-rejections=strict -r ts-node/register ./src/setup.ts",
-    "start:watch": "nodemon"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/MrMasterplan/rwklevelfiles.git"
-  },
-  "author": "",
-  "license": "GPL-3.0-or-later",
-  "bugs": {
-    "url": "https://github.com/MrMasterplan/rwklevelfiles/issues"
-  },
-  "homepage": "https://github.com/MrMasterplan/rwklevelfiles#readme",
-  "devDependencies": {
-    "@types/fs-extra": "^9.0.11",
-    "@types/node": "^14.14.37",
-    "tiled-types": "^1.3.0",
-    "ts-loader": "^9.1.0",
-    "ts-node": "^9.1.1",
-    "ts-node-dev": "^1.1.6",
-    "typescript": "^4.2.4",
-    "webpack": "^5.35.0",
-    "webpack-cli": "^4.6.0"
-  },
-  "dependencies": {
-    "fs-extra": "^9.1.0"
-  }
-}
