--- tiled/tsconfig.json
@@ -1,25 +1,0 @@
-{
-  "compilerOptions": {
-    //"lib": [ "es2015" ],
-    "target": "es5",                          /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', 'ES2018', 'ES2019', 'ES2020', or 'ESNEXT'. */
-    "module": "es6",                     /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', 'es2020', or 'ESNext'. */
-    "typeRoots": [
-      "./node_modules/@types",
-      "./src/types"
-    ],
-    "moduleResolution": "Node",
-    "outDir": "./build",                        /* Redirect output structure to the directory. */
-//    "outFile": "./build/build.js",
-    "strict": true,                           /* Enable all strict type-checking options. */
-//    "esModuleInterop": true,                  /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
-    "forceConsistentCasingInFileNames": true  /* Disallow inconsistently-cased references to the same file. */
-  },
-  "include": [
-    "./src/**/*"
-  ],
-  "exclude": [
-    "node_modules",
-    "./src/setup.ts",
-    "tests"
-  ]
-}
